feat: day04 게시글 CRUD 리팩토링
