--- src/main/java/org/example/study/ssafystudy/auth/controller/dto/LoginRequest.java
@@ -1,9 +1,0 @@
-package org.example.study.ssafystudy.auth.controller.dto;
-
-public record LoginRequest(
-        String username,
-        String password
-
-) {
-
-}
