+++ src/main/java/org/example/study/ssafystudy/domain/auth/controller/dto/LoginRequest.java
@@ -1,0 +1,9 @@
+package org.example.study.ssafystudy.domain.auth.controller.dto;
+
+public record LoginRequest(
+        String username,
+        String password
+
+) {
+
+}
