+++ src/main/java/org/example/study/ssafystudy/domain/auth/controller/dto/LoginResponse.java
@@ -1,0 +1,11 @@
+package org.example.study.ssafystudy.domain.auth.controller.dto;
+
+public record LoginResponse(
+        String accessToken,
+        String TokenType
+) {
+
+    public static LoginResponse from(String accessToken){
+        return new LoginResponse(accessToken, "Bearer ");
+    }
+}
