+++ src/main/java/org/example/study/ssafystudy/domain/member/controller/dto/MemberResponse.java
@@ -1,0 +1,18 @@
+package org.example.study.ssafystudy.domain.member.controller.dto;
+
+import org.example.study.ssafystudy.domain.member.entity.MemberEntity;
+
+public record MemberResponse(
+        Long id,
+        String username,
+        String nickname
+) {
+
+    public static MemberResponse from(MemberEntity member) {
+        return new MemberResponse(
+                member.getId(),
+                member.getUsername(),
+                member.getNickname()
+        );
+    }
+}
