--- src/main/java/org/example/study/ssafystudy/member/controller/dto/MemberRequest.java
@@ -1,18 +1,0 @@
-package org.example.study.ssafystudy.member.controller.dto;
-
-import org.example.study.ssafystudy.member.entity.MemberEntity;
-
-public record MemberRequest(
-        String username,
-        String password,
-        String nickname) {
-
-
-    public MemberEntity toEntity(){
-        return MemberEntity.create(username, password, nickname);
-    }
-
-}
-
-
-
