--- src/main/java/org/example/study/ssafystudy/member/controller/dto/MemberResponse.java
@@ -1,18 +1,0 @@
-package org.example.study.ssafystudy.member.controller.dto;
-
-import org.example.study.ssafystudy.member.entity.MemberEntity;
-
-public record MemberResponse(
-        Long id,
-        String username,
-        String nickname
-) {
-
-    public static MemberResponse from(MemberEntity member) {
-        return new MemberResponse(
-                member.getId(),
-                member.getUsername(),
-                member.getNickname()
-        );
-    }
-}
